pivoting: Setup a permutation vector
Create a permutation vector which holds the original global row numbers
as they are before LU starts, and permute its values as pivoting proceeds.
The final vector at the end of the factorization should reflect the effects
of the complete sequence of pivoting operations and indicate which row ended
where.