Fixed pivot tags on "sendPivotData"
sendPivotData was being tagged with the row where the maximum value
was found instead of the row where it was to be swapped (active row at
that time). If the maximum value was found in the same row twice, this
would cause the tag to not be unique.

Change the tag to the active row in all cases, which should always be
unique.