Enable pivot batch sizes < BLKSIZE
The diagonal chares can now send out pivot operations in multiple batches *during*
the factorization instead of in a single batch at the end of every processLU().

The batch size is currently a compile-time constant set to BLKSIZE/4.
Also, the batch size need not be the only factor determining the frequency
of pivot sends. There is quite a bit of room for experimentation here.