Fix the priorities to be based on chare column index (thisIndex.y) and not row idx.
- The computeU on the activeRow and the trailing submatrix updates should now be
  prioritized based on the column index of the chare and not the row index.
- This should prevent computation for later dependencies from interfering with the
  critical path and actually let us see the overlap of multiple LU steps.

We have thus far always seen all the trailing updates for a row getting processed
immediately after the LU of that diagonal chare. We do *NOT* want this. Ideally,
we should see the next LU getting triggered and the trailing updates moving gracefully
out of the way (as they're not on the critical path). This will help us overlap these
trailing updates with processLUs for later steps.