Modify memory resend method so it receives L and U at any time.
Previously, the memory resend mode was engaged once a block finished
the trailing pivots and updates. However, this is too late because the
memory may have alreadye exceeded the threshold, causing the program
to crash. By moving the code that receives U and L outside of SDAG,
these messages are maintained asynchronously.