Record and agglomerate pivots before performing the diagonal update
This might allow other chares to start their pivoting operations when the diagonal
chare is performing its diagonal update. However its not guaranteed to improve
performance. Perhaps worth experimenting with.

A bigger motivation comes from the simple fact that we want to record the current
pivot before we compute multipliers and find the next pivot :).