Multicast Usegment directly from its original owner instead of via diagonal chare
In cases where the pivot row is not on the diagonal chare, this will help get the
Usegment to all active panel chares sooner. The lower latency should boost performance.

earlier:
   original pivot row owner -> diag chare
   diag chare -> active panel

now:
   original pivot row owner -> active panel

The diagonal chare is not included in this active panel section. But it gets the whole
pivot row (not just the U segment) in a pivot swap msg. Hence, it does not have to be
included in this multicast.
This can change if we ever implement dont-pivot-L. That way, the pivot swap msgs will
become identical to the U segment msg, and we can reduce one point-2-point msg in the
active panel comm.