Point-to-point pivot send from diagonal to active panel chare.
Instead of multicasting the doPivot message to the entire active
panel, sent a single message to the chare that has the maximum
element.

doPivot is not longer needed. Instead, the row is sent directory.

Re-introduce scheduling of the update instead of waiting for pivot.

activePanel not longer includes the owning block (this was the case
before for the diagonal anyway). So extend this idea to the others.