Strong scaling: Schedule sub-block update after pivot redn
Traces for 1K cores of BGP indicate that the pivot reduction is taking longer
because the sub-block update occupies the PEs. Thus the reduction takes
too long to complete leading to a delay in the next column's factorization, and
hence significant idle time on this critical path computation.

Basically, we need a better mechanism of overlapping this communication with
the computation. charm's typical approaches dont help because there are no
other entry methods to be executed on these cores during this phase (active
panel) of the factorization.

Solution being explored is to actually schedule the sub-block updates after
this critical reduction. This is done by registering event listeners with the
reduction library so that the sub-block updates on a given PE are scheduled
only when the reduction has crossed that PE.